+++ CHANGELOG.md
@@ -21,10 +21,12 @@ For more information on SemVer, please visit http://semver.org/.
 ---
 
 ## Version 1.4.8
-### 2013-12-??
+### 2014-01-??
 
 #### Enhancements
 
+* Added `WPDKUser::getTransientWithUser()`
+* Added `WPDKUser::getTransient()`
 * Improved menu divider color for WordPress 3.8 admin themes scheme
 * Added Dutch localization by Frans Pronk beheer@ifra.nl
 
@@ -32,6 +34,11 @@ For more information on SemVer, please visit http://semver.org/.
 
 * Fixed set current user hooks
 * Fixed WordPress 3.8 menu
+
+#### Deprecated
+
+* `wpdk_get_user_transient()` use WPDKUser::getTransientWithUser()` instead
+
 
 ## Version 1.4.7
 ### 2013-12-17